+++ packages/sdk/src/contract-functions/index.ts
@@ -19,8 +19,6 @@ export * from './unstakeLBTC';
 // VAULT READ functions:
 export * from './getBtceShares';
 export * from './getEarnPosition';
-export * from './getSharesByAddress';
-export * from './getShareValue';
 
 // VAULT WRITE functions:
 export * from './cancelEarnWithdrawal';